Update to 21w39a
This is the update that includes support for loading
old worlds, but does not support blending them.
Nonetheless, the DataFixers are important to get right.

There are two critical issues that I discovered in Vanilla
while testing this update:
- https://bugs.mojang.com/browse/MC-238076
- https://bugs.mojang.com/browse/MC-238056

The first is fixed properly in DataConverter. Mojang will likely
do the same fix.

However, the fix for the 2nd issue is egregious. I've sanitised
the palette and data to ensure no duplicates exist. It makes an
assumption about block data: It must have minimum bits per object
4 written to disk, which is true for Vanilla, but I don't know if this
is true for all data. If Mojang applies the proper fix, which is to
accept duplicate entries in the palette, this fix can thankfully
be nuked off the planet.